LayoutTests/http/tests/contentextensions/popups-expected.txt

ALERT: about:blank successfully blocked
ALERT: data url should not be blocked
ALERT: data url should not be blocked
